Integrate voice cloning into generate voice page with enhanced features
