+++ src/components/dashboard/GenerateVoice.tsx
@@ -1,12 +1,30 @@
 
-import { useState, useEffect } from "react";
+import { useState, useEffect, useRef, useCallback } from "react";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Textarea } from "@/components/ui/textarea";
 import { Label } from "@/components/ui/label";
+import { Input } from "@/components/ui/input";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
 import { Slider } from "@/components/ui/slider";
-import { Play, Pause, Download, Volume2, Upload, FileText, RefreshCw } from "lucide-react";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { Progress } from "@/components/ui/progress";
+import { Alert, AlertDescription } from "@/components/ui/alert";
+import { Badge } from "@/components/ui/badge";
+import { 
+  Play, 
+  Pause, 
+  Download, 
+  Volume2, 
+  Upload, 
+  FileText, 
+  RefreshCw, 
+  Brain,
+  FileAudio,
+  Loader2,
+  CheckCircle2,
+  Zap
+} from "lucide-react";
 import { supabase } from "@/integrations/supabase/client";
 import { useAuth } from "@/hooks/useAuth";
 import { useToast } from "@/hooks/use-toast";
@@ -17,11 +35,32 @@ interface Voice {
   language: string;
   voice_type: string;
   description?: string;
+  external_voice_id?: string;
+  provider?: string;
+}
+
+interface VoiceSettings {
+  stability: number;
+  similarity_boost: number;
+  style: number;
+  use_speaker_boost: boolean;
+  model_id: string;
+}
+
+interface ClonedVoice {
+  id: string;
+  name: string;
+  description: string;
+  external_voice_id: string;
+  provider: string;
+  created_at: string;
 }
 
 const GenerateVoice = () => {
   const { user } = useAuth();
   const { toast } = useToast();
+  
+  // Original generation states
   const [voices, setVoices] = useState<Voice[]>([]);
   const [selectedVoice, setSelectedVoice] = useState("");
   const [text, setText] = useState("");
@@ -35,6 +74,30 @@ const GenerateVoice = () => {
   const [transliteratedText, setTransliteratedText] = useState("");
   const [isTransliterating, setIsTransliterating] = useState(false);
   const [recentGenerations, setRecentGenerations] = useState<any[]>([]);
+
+  // Voice cloning states
+  const [isCloning, setIsCloning] = useState(false);
+  const [cloneProgress, setCloneProgress] = useState(0);
+  const [selectedFile, setSelectedFile] = useState<File | null>(null);
+  const [voiceName, setVoiceName] = useState("");
+  const [voiceDescription, setVoiceDescription] = useState("");
+  const [isDragOver, setIsDragOver] = useState(false);
+  const [clonedVoices, setClonedVoices] = useState<ClonedVoice[]>([]);
+  const [usage, setUsage] = useState<any>(null);
+
+  // Voice settings for cloning
+  const [voiceSettings, setVoiceSettings] = useState<VoiceSettings>({
+    stability: 0.5,
+    similarity_boost: 0.8,
+    style: 0.0,
+    use_speaker_boost: true,
+    model_id: 'eleven_multilingual_v2'
+  });
+
+  // Refs
+  const fileInputRef = useRef<HTMLInputElement>(null);
+  const audioRef = useRef<HTMLAudioElement>(null);
+  const dropRef = useRef<HTMLDivElement>(null);
 
   // Language mapping for transliteration
   const languageMap: { [key: string]: string } = {
@@ -77,48 +140,192 @@ const GenerateVoice = () => {
     { value: "cheerful", label: "Cheerful" }
   ];
 
+  // Handle file selection for cloning
+  const handleFileSelect = useCallback((files: FileList | null) => {
+    if (!files || files.length === 0) return;
+    
+    const file = files[0];
+    const maxSize = 25 * 1024 * 1024; // 25MB
+    
+    if (file.size > maxSize) {
+      toast({
+        title: "File too large",
+        description: "Please select an audio file smaller than 25MB",
+        variant: "destructive"
+      });
+      return;
+    }
+    
+    const allowedTypes = ['audio/mpeg', 'audio/wav', 'audio/mp3', 'audio/m4a', 'audio/ogg', 'audio/webm'];
+    if (!allowedTypes.some(type => file.type.includes(type.split('/')[1]) || file.name.toLowerCase().includes(type.split('/')[1]))) {
+      toast({
+        title: "Invalid file type",
+        description: "Please select an audio file (MP3, WAV, M4A, OGG, WebM)",
+        variant: "destructive"
+      });
+      return;
+    }
+    
+    setSelectedFile(file);
+  }, [toast]);
+
+  // Drag and drop handlers
+  const handleDragOver = useCallback((e: React.DragEvent) => {
+    e.preventDefault();
+    setIsDragOver(true);
+  }, []);
+
+  const handleDragLeave = useCallback((e: React.DragEvent) => {
+    e.preventDefault();
+    setIsDragOver(false);
+  }, []);
+
+  const handleDrop = useCallback((e: React.DragEvent) => {
+    e.preventDefault();
+    setIsDragOver(false);
+    handleFileSelect(e.dataTransfer.files);
+  }, [handleFileSelect]);
+
+  // Clone voice function
+  const cloneVoice = async () => {
+    if (!selectedFile || !voiceName.trim() || !user) {
+      toast({
+        title: "Missing information",
+        description: "Please select an audio file and enter a voice name",
+        variant: "destructive"
+      });
+      return;
+    }
+
+    setIsCloning(true);
+    setCloneProgress(0);
+
+    try {
+      const formData = new FormData();
+      formData.append('audio', selectedFile);
+      formData.append('name', voiceName.trim());
+      formData.append('description', voiceDescription.trim());
+      formData.append('user_id', user.id);
+      formData.append('language', 'en-US');
+      formData.append('category', 'cloned');
+
+      // Simulate progress
+      const progressInterval = setInterval(() => {
+        setCloneProgress(prev => Math.min(prev + 10, 90));
+      }, 500);
+
+      const response = await fetch('/api/voice-cloning/clone', {
+        method: 'POST',
+        body: formData,
+      });
+
+      clearInterval(progressInterval);
+      setCloneProgress(100);
+
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error.details || 'Failed to clone voice');
+      }
+
+      const result = await response.json();
+
+      toast({
+        title: "Voice cloned successfully!",
+        description: `${voiceName} is now available for text-to-speech generation`,
+      });
+
+      // Reset form
+      setSelectedFile(null);
+      setVoiceName("");
+      setVoiceDescription("");
+      setCloneProgress(0);
+      
+      // Refresh voices list
+      await Promise.all([loadClonedVoices(), fetchData()]);
+      
+    } catch (error) {
+      console.error('Voice cloning error:', error);
+      toast({
+        title: "Voice cloning failed",
+        description: error instanceof Error ? error.message : "An unexpected error occurred",
+        variant: "destructive"
+      });
+    } finally {
+      setIsCloning(false);
+      setCloneProgress(0);
+    }
+  };
+
+  // Load cloned voices
+  const loadClonedVoices = async () => {
+    try {
+      const response = await fetch('/api/voices?user_id=' + user?.id);
+      if (response.ok) {
+        const data = await response.json();
+        setClonedVoices(data.data.filter((v: any) => v.provider === 'elevenlabs'));
+      }
+    } catch (error) {
+      console.error('Error loading cloned voices:', error);
+    }
+  };
+
+  // Load usage info
+  const loadUsage = async () => {
+    try {
+      const response = await fetch('/api/voice-cloning/usage');
+      if (response.ok) {
+        const data = await response.json();
+        setUsage(data.data);
+      }
+    } catch (error) {
+      console.error('Error loading usage:', error);
+    }
+  };
+
   // Fetch user's voices and recent generations from Supabase
+  const fetchData = async () => {
+    if (!user) return;
+
+    // Fetch voices
+    const { data: voicesData, error: voicesError } = await supabase
+      .from('voices')
+      .select('*')
+      .eq('user_id', user.id)
+      .order('created_at', { ascending: false });
+
+    if (voicesError) {
+      console.error('Error fetching voices:', voicesError);
+      toast({
+        title: "Error",
+        description: "Failed to load your voices.",
+        variant: "destructive"
+      });
+    } else {
+      setVoices(voicesData || []);
+    }
+
+    // Fetch recent generations
+    const { data: generationsData, error: generationsError } = await supabase
+      .from('generated_voices')
+      .select(`
+        *,
+        voices:voice_id (name, voice_type)
+      `)
+      .eq('user_id', user.id)
+      .order('created_at', { ascending: false })
+      .limit(5);
+
+    if (generationsError) {
+      console.error('Error fetching generations:', generationsError);
+    } else {
+      setRecentGenerations(generationsData || []);
+    }
+  };
+
   useEffect(() => {
-    const fetchData = async () => {
-      if (!user) return;
-
-      // Fetch voices
-      const { data: voicesData, error: voicesError } = await supabase
-        .from('voices')
-        .select('*')
-        .eq('user_id', user.id)
-        .order('created_at', { ascending: false });
-
-      if (voicesError) {
-        console.error('Error fetching voices:', voicesError);
-        toast({
-          title: "Error",
-          description: "Failed to load your voices.",
-          variant: "destructive"
-        });
-      } else {
-        setVoices(voicesData || []);
-      }
-
-      // Fetch recent generations
-      const { data: generationsData, error: generationsError } = await supabase
-        .from('generated_voices')
-        .select(`
-          *,
-          voices:voice_id (name, voice_type)
-        `)
-        .eq('user_id', user.id)
-        .order('created_at', { ascending: false })
-        .limit(5);
-
-      if (generationsError) {
-        console.error('Error fetching generations:', generationsError);
-      } else {
-        setRecentGenerations(generationsData || []);
-      }
-    };
-
-    fetchData();
+    if (user) {
+      Promise.all([fetchData(), loadClonedVoices(), loadUsage()]);
+    }
   }, [user, toast]);
 
   // Enhanced Google Transliteration API integration
@@ -237,6 +444,10 @@ const GenerateVoice = () => {
           setText(content);
         };
         reader.readAsText(file);
+        toast({
+          title: "File uploaded successfully",
+          description: `Loaded content from ${file.name}`,
+        });
       } else if (file.type === 'application/pdf' || file.name.endsWith('.pdf')) {
         // Basic PDF text extraction using pdfjs-dist would go here
         toast({
@@ -335,15 +546,67 @@ const GenerateVoice = () => {
     }
   };
 
-  const handlePlayPause = () => {
-    setIsPlaying(!isPlaying);
-    // In production, implement actual audio playback
-    console.log("Playing generated audio:", !isPlaying);
-    
-    if (!isPlaying) {
-      // Simulate playback
-      setTimeout(() => setIsPlaying(false), 5000);
+  // Generate speech function for cloned voices
+  const generateSpeech = async () => {
+    if (!text.trim() || !selectedVoice) {
+      toast({
+        title: "Missing information",
+        description: "Please enter text and select a voice",
+        variant: "destructive"
+      });
+      return;
     }
+
+    setIsGenerating(true);
+
+    try {
+      const response = await fetch('/api/voice-cloning/synthesize', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          text: transliteratedText || text,
+          voice_id: selectedVoice,
+          options: voiceSettings,
+        }),
+      });
+
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error.details || 'Failed to generate speech');
+      }
+
+      const audioBlob = await response.blob();
+      const audioUrl = URL.createObjectURL(audioBlob);
+      setGeneratedAudio(audioUrl);
+
+      toast({
+        title: "Speech generated successfully!",
+        description: "Your audio is ready to play and download",
+      });
+
+    } catch (error) {
+      console.error('Speech generation error:', error);
+      toast({
+        title: "Speech generation failed",
+        description: error instanceof Error ? error.message : "An unexpected error occurred",
+        variant: "destructive"
+      });
+    } finally {
+      setIsGenerating(false);
+    }
+  };
+
+  const handlePlayPause = () => {
+    if (!audioRef.current || !generatedAudio) return;
+    
+    if (isPlaying) {
+      audioRef.current.pause();
+    } else {
+      audioRef.current.play();
+    }
+    setIsPlaying(!isPlaying);
   };
 
   const handleDownload = async () => {
@@ -357,34 +620,18 @@ const GenerateVoice = () => {
     }
 
     try {
-      toast({
-        title: "Download starting...",
-        description: "Preparing generated audio for download",
-      });
-
-      // In production, this would be the actual generated audio URL
-      // For now, create a mock audio file blob
-      const mockAudioBlob = new Blob(
-        [new ArrayBuffer(1024 * 1024)], // 1MB mock audio file
-        { type: 'audio/wav' }
-      );
-      
-      const url = window.URL.createObjectURL(mockAudioBlob);
       const link = document.createElement('a');
-      link.href = url;
+      link.href = generatedAudio;
       
       // Create filename with selected voice and timestamp
       const selectedVoiceData = voices.find(v => v.id === selectedVoice);
       const timestamp = new Date().toISOString().slice(0, 19).replace(/:/g, '-');
-      const fileName = `generated_${selectedVoiceData?.name || 'voice'}_${timestamp}.wav`;
+      const fileName = `generated_${selectedVoiceData?.name || 'voice'}_${timestamp}.mp3`;
       link.download = fileName;
       
       document.body.appendChild(link);
       link.click();
       document.body.removeChild(link);
-      
-      // Clean up the blob URL
-      window.URL.revokeObjectURL(url);
       
       toast({
         title: "Download complete! 📁",
@@ -401,197 +648,372 @@ const GenerateVoice = () => {
   };
 
   return (
-    <div className="max-w-4xl mx-auto space-y-6">
-      {/* Input Section */}
-      <div className="grid gap-6 md:grid-cols-2">
-        {/* Text Input */}
-        <Card>
-          <CardHeader>
-            <CardTitle>Text to Speech</CardTitle>
-            <CardDescription>
-              Enter text or upload a file to convert to speech in Indian languages
-            </CardDescription>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            <div className="flex items-center space-x-2 mb-4">
-              <label className="cursor-pointer">
-                <input
-                  type="file"
-                  accept=".txt,.pdf"
-                  onChange={handleFileUpload}
-                  className="hidden"
-                />
-                <Button type="button" variant="outline" size="sm">
-                  <Upload className="h-4 w-4 mr-2" />
-                  Upload File
-                </Button>
-              </label>
-              <span className="text-xs text-muted-foreground">TXT or PDF</span>
-            </div>
-            
-            <div className="space-y-2">
-              <Label htmlFor="text">Text Content</Label>
-              <Textarea
-                id="text"
-                placeholder="Enter your text here or upload a file..."
-                value={text}
-                onChange={(e) => setText(e.target.value)}
-                rows={8}
-                className="resize-none"
-              />
-              <div className="text-xs text-muted-foreground text-right">
-                {text.length} characters
-              </div>
-            </div>
+    <div className="max-w-6xl mx-auto space-y-6">
+      {/* Usage Stats */}
+      {usage && (
+        <Alert>
+          <Zap className="h-4 w-4" />
+          <AlertDescription>
+            ElevenLabs Usage: {usage.character_count?.toLocaleString() || 0} / {usage.character_limit?.toLocaleString() || 0} characters used
+          </AlertDescription>
+        </Alert>
+      )}
 
-            {/* Transliteration Panel */}
-            {text.trim() && (
-              <div className="space-y-2 mt-4 p-3 bg-muted/50 rounded-lg">
-                <div className="flex items-center justify-between">
-                  <Label className="text-sm font-semibold">Transliterated Output</Label>
-                  <Button 
-                    variant="ghost" 
-                    size="sm" 
-                    onClick={() => transliterateWithGoogle(text, outputLanguage)}
-                    disabled={isTransliterating}
-                    className="h-6 px-2"
-                  >
-                    <RefreshCw className={`h-3 w-3 ${isTransliterating ? 'animate-spin' : ''}`} />
-                  </Button>
+      <Tabs defaultValue="generate" className="w-full">
+        <TabsList className="grid w-full grid-cols-2">
+          <TabsTrigger value="generate">Generate Speech</TabsTrigger>
+          <TabsTrigger value="clone" className="flex items-center gap-2">
+            <Brain className="h-4 w-4" />
+            Clone Voice
+          </TabsTrigger>
+        </TabsList>
+
+        {/* Speech Generation Tab */}
+        <TabsContent value="generate" className="space-y-6">
+          {/* Input Section */}
+          <div className="grid gap-6 md:grid-cols-2">
+            {/* Text Input */}
+            <Card>
+              <CardHeader>
+                <CardTitle>Text to Speech</CardTitle>
+                <CardDescription>
+                  Enter text or upload a file to convert to speech in Indian languages
+                </CardDescription>
+              </CardHeader>
+              <CardContent className="space-y-4">
+                <div className="flex items-center space-x-2 mb-4">
+                  <label className="cursor-pointer">
+                    <input
+                      type="file"
+                      accept=".txt,.pdf"
+                      onChange={handleFileUpload}
+                      className="hidden"
+                    />
+                    <Button type="button" variant="outline" size="sm">
+                      <Upload className="h-4 w-4 mr-2" />
+                      Upload File
+                    </Button>
+                  </label>
+                  <span className="text-xs text-muted-foreground">TXT or PDF</span>
                 </div>
-                <div className="text-sm text-muted-foreground min-h-[60px] p-2 bg-background rounded border">
-                  {isTransliterating ? (
-                    <div className="flex items-center justify-center">
-                      <RefreshCw className="h-4 w-4 animate-spin mr-2" />
-                      Transliterating with Google API...
+                
+                <div className="space-y-2">
+                  <Label htmlFor="text">Text Content</Label>
+                  <Textarea
+                    id="text"
+                    placeholder="Enter your text here or upload a file..."
+                    value={text}
+                    onChange={(e) => setText(e.target.value)}
+                    rows={8}
+                    className="resize-none"
+                  />
+                  <div className="text-xs text-muted-foreground text-right">
+                    {text.length} characters
+                  </div>
+                </div>
+
+                {/* Transliteration Panel */}
+                {text.trim() && (
+                  <div className="space-y-2 mt-4 p-3 bg-muted/50 rounded-lg">
+                    <div className="flex items-center justify-between">
+                      <Label className="text-sm font-semibold">Transliterated Output</Label>
+                      <Button 
+                        variant="ghost" 
+                        size="sm" 
+                        onClick={() => transliterateWithGoogle(text, outputLanguage)}
+                        disabled={isTransliterating}
+                        className="h-6 px-2"
+                      >
+                        <RefreshCw className={`h-3 w-3 ${isTransliterating ? 'animate-spin' : ''}`} />
+                      </Button>
                     </div>
-                  ) : (
-                    <div className="space-y-1">
-                      <div className="font-medium">{transliteratedText || "Enter text to see transliteration"}</div>
-                      {transliteratedText && (
-                        <div className="text-xs text-blue-600">
-                          ✓ Ready for voice generation
+                    <div className="text-sm text-muted-foreground min-h-[60px] p-2 bg-background rounded border">
+                      {isTransliterating ? (
+                        <div className="flex items-center justify-center">
+                          <RefreshCw className="h-4 w-4 animate-spin mr-2" />
+                          Transliterating with Google API...
+                        </div>
+                      ) : (
+                        <div className="space-y-1">
+                          <div className="font-medium">{transliteratedText || "Enter text to see transliteration"}</div>
+                          {transliteratedText && (
+                            <div className="text-xs text-blue-600">
+                              ✓ Ready for voice generation
+                            </div>
+                          )}
                         </div>
                       )}
                     </div>
+                    <p className="text-xs text-muted-foreground">
+                      Auto-transliteration enabled • Click refresh to manually update
+                    </p>
+                  </div>
+                )}
+              </CardContent>
+            </Card>
+
+            {/* Voice Settings */}
+            <Card>
+              <CardHeader>
+                <CardTitle>Voice Settings</CardTitle>
+                <CardDescription>
+                  Configure voice parameters and quality
+                </CardDescription>
+              </CardHeader>
+              <CardContent className="space-y-6">
+                <div className="space-y-2">
+                  <Label>Select Voice</Label>
+                  <Select value={selectedVoice} onValueChange={setSelectedVoice}>
+                    <SelectTrigger>
+                      <SelectValue placeholder="Choose from your voice library" />
+                    </SelectTrigger>
+                    <SelectContent>
+                      {voices.map((voice) => (
+                        <SelectItem key={voice.id} value={voice.external_voice_id || voice.id}>
+                          <div className="flex flex-col">
+                            <span className="flex items-center gap-2">
+                              {voice.name}
+                              {voice.provider && (
+                                <Badge variant="secondary" className="text-xs">
+                                  {voice.provider}
+                                </Badge>
+                              )}
+                            </span>
+                            <span className="text-xs text-muted-foreground">
+                              {voice.voice_type} • {voice.language.toUpperCase()}
+                            </span>
+                          </div>
+                        </SelectItem>
+                      ))}
+                      {clonedVoices.map((voice) => (
+                        <SelectItem key={voice.id} value={voice.external_voice_id}>
+                          <div className="flex flex-col">
+                            <span className="flex items-center gap-2">
+                              {voice.name}
+                              <Badge variant="secondary" className="text-xs">
+                                cloned
+                              </Badge>
+                            </span>
+                            <span className="text-xs text-muted-foreground">
+                              {voice.description}
+                            </span>
+                          </div>
+                        </SelectItem>
+                      ))}
+                    </SelectContent>
+                  </Select>
+                  {voices.length === 0 && clonedVoices.length === 0 && (
+                    <p className="text-xs text-muted-foreground">
+                      No voices found. Add voices in the "Add Voice" tab or clone a voice first.
+                    </p>
                   )}
                 </div>
-                <p className="text-xs text-muted-foreground">
-                  Auto-transliteration enabled • Click refresh to manually update
-                </p>
+
+                <div className="space-y-2">
+                  <Label>Output Language</Label>
+                  <Select value={outputLanguage} onValueChange={setOutputLanguage}>
+                    <SelectTrigger>
+                      <SelectValue />
+                    </SelectTrigger>
+                    <SelectContent>
+                      {indianLanguages.map((lang) => (
+                        <SelectItem key={lang.value} value={lang.value}>
+                          {lang.label}
+                        </SelectItem>
+                      ))}
+                    </SelectContent>
+                  </Select>
+                </div>
+
+                <div className="space-y-3">
+                  <Label>Speed: {speed[0].toFixed(1)}x</Label>
+                  <Slider
+                    value={speed}
+                    onValueChange={setSpeed}
+                    max={2}
+                    min={0.5}
+                    step={0.1}
+                    className="w-full"
+                  />
+                </div>
+
+                <div className="space-y-3">
+                  <Label>Pitch: {pitch[0].toFixed(1)}x</Label>
+                  <Slider
+                    value={pitch}
+                    onValueChange={setPitch}
+                    max={2}
+                    min={0.5}
+                    step={0.1}
+                    className="w-full"
+                  />
+                </div>
+
+                <div className="space-y-2">
+                  <Label>Tone</Label>
+                  <Select value={tone} onValueChange={setTone}>
+                    <SelectTrigger>
+                      <SelectValue />
+                    </SelectTrigger>
+                    <SelectContent>
+                      {tones.map((t) => (
+                        <SelectItem key={t.value} value={t.value}>
+                          {t.label}
+                        </SelectItem>
+                      ))}
+                    </SelectContent>
+                  </Select>
+                </div>
+
+                <Button 
+                  onClick={handleGenerate}
+                  className="w-full"
+                  disabled={isGenerating || !selectedVoice || !text.trim()}
+                >
+                  {isGenerating ? (
+                    <>
+                      <Loader2 className="h-4 w-4 mr-2 animate-spin" />
+                      Generating...
+                    </>
+                  ) : (
+                    <>
+                      <Volume2 className="h-4 w-4 mr-2" />
+                      Generate Speech
+                    </>
+                  )}
+                </Button>
+              </CardContent>
+            </Card>
+          </div>
+        </TabsContent>
+
+        {/* Voice Cloning Tab */}
+        <TabsContent value="clone" className="space-y-6">
+          <Card>
+            <CardHeader>
+              <CardTitle className="flex items-center gap-2">
+                <Brain className="h-5 w-5" />
+                Clone Your Voice
+              </CardTitle>
+              <CardDescription>
+                Upload a 30-second audio sample to create a custom voice clone
+              </CardDescription>
+            </CardHeader>
+            <CardContent className="space-y-4">
+              {/* File Upload Area */}
+              <div
+                ref={dropRef}
+                className={`border-2 border-dashed rounded-lg p-8 text-center transition-colors ${
+                  isDragOver ? 'border-primary bg-primary/5' : 'border-muted-foreground/25'
+                }`}
+                onDragOver={handleDragOver}
+                onDragLeave={handleDragLeave}
+                onDrop={handleDrop}
+              >
+                <div className="flex flex-col items-center gap-4">
+                  <FileAudio className="h-12 w-12 text-muted-foreground" />
+                  {selectedFile ? (
+                    <div className="space-y-2">
+                      <p className="font-medium text-green-600">
+                        Selected: {selectedFile.name}
+                      </p>
+                      <p className="text-sm text-muted-foreground">
+                        Size: {(selectedFile.size / (1024 * 1024)).toFixed(2)} MB
+                      </p>
+                    </div>
+                  ) : (
+                    <>
+                      <p className="text-lg font-medium">
+                        Drop audio file here or click to browse
+                      </p>
+                      <p className="text-sm text-muted-foreground">
+                        Supports MP3, WAV, M4A, OGG, WebM (max 25MB)
+                      </p>
+                    </>
+                  )}
+                  <Button
+                    variant="outline"
+                    onClick={() => fileInputRef.current?.click()}
+                    disabled={isCloning}
+                  >
+                    <Upload className="h-4 w-4 mr-2" />
+                    Choose File
+                  </Button>
+                  <input
+                    ref={fileInputRef}
+                    type="file"
+                    accept="audio/*"
+                    className="hidden"
+                    onChange={(e) => handleFileSelect(e.target.files)}
+                  />
+                </div>
               </div>
-            )}
-          </CardContent>
-        </Card>
 
-        {/* Voice Settings */}
-        <Card>
-          <CardHeader>
-            <CardTitle>Voice Settings</CardTitle>
-            <CardDescription>
-              Configure voice parameters and quality
-            </CardDescription>
-          </CardHeader>
-          <CardContent className="space-y-6">
-            <div className="space-y-2">
-              <Label>Select Voice</Label>
-              <Select value={selectedVoice} onValueChange={setSelectedVoice}>
-                <SelectTrigger>
-                  <SelectValue placeholder="Choose from your voice library" />
-                </SelectTrigger>
-                <SelectContent>
-                  {voices.map((voice) => (
-                    <SelectItem key={voice.id} value={voice.id}>
-                      <div className="flex flex-col">
-                        <span>{voice.name}</span>
-                        <span className="text-xs text-muted-foreground">
-                          {voice.voice_type} • {voice.language.toUpperCase()}
-                        </span>
-                      </div>
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-              {voices.length === 0 && (
-                <p className="text-xs text-muted-foreground">
-                  No voices found. Add voices in the "Add Voice" tab first.
-                </p>
+              {/* Voice Information */}
+              <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+                <div className="space-y-2">
+                  <Label htmlFor="voiceName">Voice Name *</Label>
+                  <Input
+                    id="voiceName"
+                    placeholder="e.g., My Professional Voice"
+                    value={voiceName}
+                    onChange={(e) => setVoiceName(e.target.value)}
+                    disabled={isCloning}
+                  />
+                </div>
+                <div className="space-y-2">
+                  <Label htmlFor="voiceDescription">Description</Label>
+                  <Input
+                    id="voiceDescription"
+                    placeholder="Optional description"
+                    value={voiceDescription}
+                    onChange={(e) => setVoiceDescription(e.target.value)}
+                    disabled={isCloning}
+                  />
+                </div>
+              </div>
+
+              {/* Progress Bar */}
+              {isCloning && (
+                <div className="space-y-2">
+                  <div className="flex justify-between text-sm">
+                    <span>Cloning voice...</span>
+                    <span>{cloneProgress}%</span>
+                  </div>
+                  <Progress value={cloneProgress} className="w-full" />
+                </div>
               )}
-            </div>
 
-            <div className="space-y-2">
-              <Label>Output Language</Label>
-              <Select value={outputLanguage} onValueChange={setOutputLanguage}>
-                <SelectTrigger>
-                  <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                  {indianLanguages.map((lang) => (
-                    <SelectItem key={lang.value} value={lang.value}>
-                      {lang.label}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
-
-            <div className="space-y-3">
-              <Label>Speed: {speed[0].toFixed(1)}x</Label>
-              <Slider
-                value={speed}
-                onValueChange={setSpeed}
-                max={2}
-                min={0.5}
-                step={0.1}
+              {/* Clone Button */}
+              <Button
+                onClick={cloneVoice}
+                disabled={!selectedFile || !voiceName.trim() || isCloning}
                 className="w-full"
-              />
-            </div>
-
-            <div className="space-y-3">
-              <Label>Pitch: {pitch[0].toFixed(1)}x</Label>
-              <Slider
-                value={pitch}
-                onValueChange={setPitch}
-                max={2}
-                min={0.5}
-                step={0.1}
-                className="w-full"
-              />
-            </div>
-
-            <div className="space-y-2">
-              <Label>Tone</Label>
-              <Select value={tone} onValueChange={setTone}>
-                <SelectTrigger>
-                  <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                  {tones.map((t) => (
-                    <SelectItem key={t.value} value={t.value}>
-                      {t.label}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
-
-            <Button 
-              onClick={handleGenerate}
-              className="w-full"
-              disabled={isGenerating || !selectedVoice || !text.trim()}
-            >
-              {isGenerating ? "Generating..." : "Generate Speech"}
-            </Button>
-          </CardContent>
-        </Card>
-      </div>
+                size="lg"
+              >
+                {isCloning ? (
+                  <>
+                    <Loader2 className="h-4 w-4 mr-2 animate-spin" />
+                    Cloning Voice...
+                  </>
+                ) : (
+                  <>
+                    <Brain className="h-4 w-4 mr-2" />
+                    Clone Voice
+                  </>
+                )}
+              </Button>
+            </CardContent>
+          </Card>
+        </TabsContent>
+      </Tabs>
 
       {/* Generated Audio Section */}
       {generatedAudio && (
         <Card>
           <CardHeader>
             <CardTitle className="flex items-center space-x-2">
-              <Volume2 className="h-5 w-5" />
+              <CheckCircle2 className="h-5 w-5 text-green-500" />
               <span>Generated Audio</span>
             </CardTitle>
             <CardDescription>
@@ -628,6 +1050,12 @@ const GenerateVoice = () => {
                 {isPlaying ? "Playing..." : "Ready"}
               </span>
             </div>
+            <audio
+              ref={audioRef}
+              src={generatedAudio}
+              onEnded={() => setIsPlaying(false)}
+              className="hidden"
+            />
           </CardContent>
         </Card>
       )}
